unify auto between transformers & diffusers
