+++ src/optimum/rbln/diffusers/__init__.py
@@ -59,9 +59,6 @@ _import_structure = {
         "RBLNVQModelConfig",
     ],
     "pipelines": [
-        "RBLNAutoPipelineForText2Image",
-        "RBLNAutoPipelineForImage2Image",
-        "RBLNAutoPipelineForInpainting",
         "RBLNCosmosTextToWorldPipeline",
         "RBLNCosmosVideoToWorldPipeline",
         "RBLNCosmosSafetyChecker",
@@ -147,9 +144,6 @@ if TYPE_CHECKING:
         RBLNVQModel,
     )
     from .pipelines import (
-        RBLNAutoPipelineForImage2Image,
-        RBLNAutoPipelineForInpainting,
-        RBLNAutoPipelineForText2Image,
         RBLNCosmosSafetyChecker,
         RBLNCosmosTextToWorldPipeline,
         RBLNCosmosVideoToWorldPipeline,
