+++ src/optimum/rbln/transformers/__init__.py
@@ -21,20 +21,6 @@ _import_structure = {
     "models": [
         "RBLNASTForAudioClassification",
         "RBLNASTForAudioClassificationConfig",
-        "RBLNAutoModel",
-        "RBLNAutoModelForAudioClassification",
-        "RBLNAutoModelForCausalLM",
-        "RBLNAutoModelForCTC",
-        "RBLNAutoModelForDepthEstimation",
-        "RBLNAutoModelForImageClassification",
-        "RBLNAutoModelForImageTextToText",
-        "RBLNAutoModelForMaskedLM",
-        "RBLNAutoModelForQuestionAnswering",
-        "RBLNAutoModelForSeq2SeqLM",
-        "RBLNAutoModelForSequenceClassification",
-        "RBLNAutoModelForSpeechSeq2Seq",
-        "RBLNAutoModelForVision2Seq",
-        "RBLNAutoModelForTextEncoding",
         "RBLNBartForConditionalGeneration",
         "RBLNBartForConditionalGenerationConfig",
         "RBLNBartModel",
@@ -160,20 +146,6 @@ if TYPE_CHECKING:
     from .models import (
         RBLNASTForAudioClassification,
         RBLNASTForAudioClassificationConfig,
-        RBLNAutoModel,
-        RBLNAutoModelForAudioClassification,
-        RBLNAutoModelForCausalLM,
-        RBLNAutoModelForCTC,
-        RBLNAutoModelForDepthEstimation,
-        RBLNAutoModelForImageClassification,
-        RBLNAutoModelForImageTextToText,
-        RBLNAutoModelForMaskedLM,
-        RBLNAutoModelForQuestionAnswering,
-        RBLNAutoModelForSeq2SeqLM,
-        RBLNAutoModelForSequenceClassification,
-        RBLNAutoModelForSpeechSeq2Seq,
-        RBLNAutoModelForTextEncoding,
-        RBLNAutoModelForVision2Seq,
         RBLNBartForConditionalGeneration,
         RBLNBartForConditionalGenerationConfig,
         RBLNBartModel,
